+++ MD/Bounds.cpp
@@ -13,7 +13,7 @@ bool Bounds::atomInBounds(Atom *a) {
 	return true;
 }
 
-float Bounds::volume() {
+double Bounds::volume() {
 	return trace.prod();
 }
